refactor(checkbox): change name from `PCheckBox` to `PCheckbox`
