+++ src/components.ts
@@ -24,8 +24,8 @@ export { default as PBoard } from './data-display/board/PBoard.vue';
 export { default as PLabel } from './data-display/label/PLabel.vue';
 export { default as PBoardItem } from './data-display/board-item/PBoardItem.vue';
 export { default as PCollapsiblePanel } from './data-display/collapsibles/collapsible-panel/PCollapsiblePanel.vue';
-export { default as PCheckBox } from './inputs/checkbox/PCheckBox.vue';
-export { default as PCheckBoxGroup } from './inputs/checkbox-group/PCheckBoxGroup.vue';
+export { default as PCheckBox } from './inputs/checkbox/PCheckbox.vue';
+export { default as PCheckBoxGroup } from './inputs/checkbox-group/PCheckboxGroup.vue';
 export { default as PFieldGroup } from './inputs/forms/field-group/PFieldGroup.vue';
 export { default as PRadio } from './inputs/radio/PRadio.vue';
 export { default as PRadioGroup } from './inputs/radio-group/PRadioGroup.vue';
