+++ src/inputs/checkbox-group/PCheckboxGroup.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Canvas, Story, ArgsTable } from '@storybook/addon-docs/blocks';
-import PCheckBoxGroup from './PCheckBoxGroup.vue';
+import PCheckBoxGroup from './PCheckboxGroup.vue';
 import { getCurrentInstance, reactive, toRefs } from 'vue';
 import { makeOptionalProxy } from '@/utils/composition-helpers';
 import { getRadioGroupArgTypes } from '@/inputs/radio-group/story-helper';
