+++ src/inputs/checkbox/PCheckBox.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Canvas, Story, ArgsTable } from '@storybook/addon-docs/blocks';
-import PCheckBox from './PCheckBox.vue';
+import PCheckBox from './PCheckbox.vue';
 import { getCurrentInstance, reactive, toRefs } from 'vue';
 import { makeOptionalProxy } from '@/utils/composition-helpers';
 import { getCheckboxArgTypes } from '@/inputs/checkbox/story-helper';
