+++ src/inputs/checkbox/PCheckbox.vue
